将图书库存数量的类型改为long long

+++ src/book.h
@@ -24,7 +24,7 @@ struct Book {
     int BookName_len = 0;
     int Author_len = 0;
     int Keyword_len{};
-    int Quantity = 0;
+    long long Quantity = 0;
     double price = 0;
     bool is_new = false;
 };
